Move tooltip config to opTooltipService
